Move py_package_version.py to metadata/
